Outsource ItemManager
